requirements.txt
python-pptx==0.6.23
Pillow==10.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0 